2dcontext/the-canvas-state/2d.state.saverestore.compareResut.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<html>
	<title>Canvas test: canvas-state</title>
	<meta name="author" content="Philip Taylor">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/common/canvas-tests.js"></script>
	<link rel="stylesheet" href="/common/canvas-tests.css">
	<style type="text/css">
		#c1,#c2,#c3,#c4{display:inline-block !important;}
	</style>
	<body class="show_output">

		<h1>canvas-state.save&restore</h1>
		<p class="desc">
			canvas-state
		</p>

		<p class="output">
			Actual output:
		</p>
		<div id="lyr_cvs">
			<canvas id="c1" class="output" width="200" height="200"></canvas>
			<canvas id="c2" class="output" width="200" height="200"></canvas>
			<canvas id="c3" class="output" width="200" height="200"></canvas>
			<canvas id="c4" class="output" width="200" height="200"></canvas>
		</div>
		<ul id="d"></ul>
		<script>

			var imgLog = [];
			var img1 = stateTest("c1",true,false);
			var img2 = stateTest("c2",true,true);
			var img3 = stateTest("c3",false,false);
			var img4 = stateTest("c4",false,true);
			test(function() {
				// get the canvas element using the DOM
				assert_equals(img1,img2,"draw rectagle in difference timming with same state");
			},"draw rectagle in difference timming with same state");
			test(function() {
				assert_not_equals(img1,img3,"draw rectagle in difference state mode(use/no use save&restore)");
			},"draw rectagle in difference state mode(use/no use save&restore)");
			test(function() {
				assert_not_equals(img2,img4,"c24");
			},"draw rectagle in difference timming and state mode(use/no use save&restore)");
			test(function() {
				assert_not_equals(img3,img4,"c34");
			},"draw rectagle in difference timming without save&restore");
			function stateTest(id, flg, flg2){
				var canvas = document.getElementById(id);

				// Make sure we don't execute when canvas isn't supported
				if (canvas.getContext) {
					
					// use getContext to use the canvas for drawing
					var ctx = canvas.getContext('2d');
					
					//  draw a rectangle with default settings
					ctx.fillStyle = '#ffdd00';
					ctx.fillRect(0,0,150,150);
					
					// Draw a rectangle with restored settings when flg2 is true
					if(flg2)ctx.fillRect(150,150,200,200);
					//  Save the default state
					if(flg) ctx.save();
					
					// Make changes to the settings
					ctx.fillStyle = '#66FFFF';
					ctx.fillRect(15,15,120,120);
					// Save the current state
					if(flg) ctx.save();
					
					// Make the new changes to the settings
					ctx.fillStyle = '#993333';
					ctx.globalAlpha = 0.5;
					ctx.fillRect(30,30,90,90);
					
					// Restore previous state
					if(flg) ctx.restore();
					// Draw a rectangle with restored settings
					ctx.fillRect(45,45,60,60);
					
					// Restore original state
					if(flg) ctx.restore();
					// Draw a rectangle with restored settings when flg2 is false
					if(!flg2) ctx.fillRect(150,150,200,200);
					
					//showDiffence(imgLog);
					return canvas.toDataURL();
				} else {
					return "";
				}
			}
		</script>
		<div id="log"></div>
	</body>
</html>
